project/requirements.txt
openai 
agno
python-dotenv
duckduckgo-search
lancedb
tantivy
pypdf
pandas
tweepy
anthropic
supabase
matplotlib
seaborn
typing
python-pptx
fastapi
mcp>=1.0.0
uvicorn>=0.32.1
langgraph
langchain
langchain-mcp-adapters
langchain-openai
pydantic_ai
langchain_community